Merged george973-ocs@8169 (partial). Bug 4416: Incorrect hex grid alignment
Says George:

Was due to HexGrid keeping some old fields (just so some old methods 
could be kept in case module own code used them).

Unfortunately the new fields which are used in calculating the grid were 
given the same names and as AutoConfigurer was being used it put the 
values in the old fields and not in the fields being used. So stored 
values never used.

Just deleting the old fields works.



git-svn-id: https://vassalengine.svn.sourceforge.net/svnroot/vassalengine/VASSAL-src/trunk@8174 67b53d14-2c14-4ace-a08f-0dab2b34000c