Merged Brent-3.2@8216.
Bug 4278 - map level global properties of the same name messed up online

Issue with the handling of any same named properties in different property cont
ainers (module, map or zone). There was no encoding of the identity of the cont
ainer when a property value was set or changed, so all references to different 
properties of teh same name ended up being handled by the first found property 
of that name.



git-svn-id: https://vassalengine.svn.sourceforge.net/svnroot/vassalengine/VASSAL-src/trunk@8218 67b53d14-2c14-4ace-a08f-0dab2b34000c