Merged Brent-3.2-New@8279. Bug 4705 -Decks behave strangely after loading saved game
When Decks are loaded from a saved game, they are not being correctly
associated with their DrawPile. This results in strange behaviour if
you use Return to Deck to return a card to a Deck it is already
supposed to be in. The card ends up unstacked in the same location,
but outside the Deck.


git-svn-id: https://vassalengine.svn.sourceforge.net/svnroot/vassalengine/VASSAL-src/trunk@8284 67b53d14-2c14-4ace-a08f-0dab2b34000c