Merged Brent-3.2-New@8282.
Fixed Bug 4696: Embellishment follow property to display first level when
property is null instead of error 

Currently, if the follow property is null, the display defaults to the
first level, however, if the follow property is blank, then an error is
displayed. Have changed the behaviour so a blank property value also
displays the first level with no error.


git-svn-id: https://vassalengine.svn.sourceforge.net/svnroot/vassalengine/VASSAL-src/trunk@8294 67b53d14-2c14-4ace-a08f-0dab2b34000c