Fixed Bug 4159:  NPE in javax.swing.BoxLayout.preferredLayoutSize()
When a component with a MigLayout has a child which is a JTextPane and
an ancestor with a BoxLayout, MigLayout calls BoxLayout.preferredLayoutSize()
recursively, which is wrong because BoxLayout is not in a consistent
state until preferredLayoutSize() completes. I've patched MigLayout not
to do this. See http://migcalendar.com/forums/viewtopic.php?p=20836#p20836
for the patches. I've used Patch #2 here.

git-svn-id: https://vassalengine.svn.sourceforge.net/svnroot/vassalengine/VASSAL-src/trunk@8473 67b53d14-2c14-4ace-a08f-0dab2b34000c